refactor: extract download functionality into own module with DownloadMenu component as entry
