+++ src/App.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import "./App.css";
 import Map from "./Search/Map";
 //import {useGeolocation} from '../src/useGeolocation';
@@ -10,27 +9,12 @@ import PlantDetail from "./Plant/PlantDetail";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faSearch,
-  faFileExcel,
   faStar,
-  faQrcode,
-  faIdCard,
 } from "@fortawesome/free-solid-svg-icons";
-import JSZip from "jszip";
-import { saveAs } from "file-saver";
-import { PDFViewer, pdf } from "@react-pdf/renderer";
+import { PDFViewer } from "@react-pdf/renderer";
 
 import Favorites from "./Favorites/Favorites";
 
-import {
-  Button,
-  Modal,
-  Col,
-  Container,
-  Row,
-  ProgressBar,
-} from "react-bootstrap";
-
-import ReactExport, { ExcelCellData, ExcelSheetData } from "react-export-excel";
 import {
   //BrowserRouter as Router,
   useLocation,
@@ -46,16 +30,12 @@ import SearchCriteriaConverter from "./Search/search-criteria-converter";
 import BenchCardDocument from "./Plant/BenchCardDocument";
 
 import {
-  BenchCardTemplate,
   BoolDict,
   City,
   Data,
-  DownloadAction,
   Plant,
   SearchCriteria,
 } from "./types";
-import { plantDetailQrCodeFromId } from "./Plant/PlantDetailQrCode";
-import DownloadActionList from './Download/DownloadActionList';
 
 interface Props {
   data: Data;
@@ -203,245 +183,6 @@ function App({ data }: Props) {
       }
     );
   }
-
-  const zipCancelled = React.useRef(false);
-  const [showZipModal, setShowZipModal] = React.useState(false);
-  const [zipCurrent, setZipCurrent] = React.useState(0);
-  const [zipTotal, setZipTotal] = React.useState(0);
-  const [currentBct, setCurrentBct] = React.useState<BenchCardTemplate | null>(
-    null
-  );
-
-  React.useEffect(() => {
-    if (showZipModal && currentBct && !zipCancelled.current) {
-      setZipTotal(favoritePlants.length);
-      setZipCurrent(0);
-      const cb = async () => {
-        try {
-          const plantBlobPairs = await Promise.all(
-            favoritePlants.map(async (p: Plant) => {
-              let b: any;
-              try {
-                if (zipCancelled.current) {
-                  throw new Error("Download cancelled");
-                }
-                b = await pdf(
-                  <BenchCardDocument
-                    benchCardTemplate={currentBct}
-                    plant={p}
-                    region={searchCriteria.city.region}
-                    waterUseByCode={data.waterUseByCode}
-                  />
-                ).toBlob();
-                let current = 0;
-                setZipCurrent((i) => {
-                  current = i === 0 ? 1 : i + 1;
-                  return current;
-                });
-                console.log(
-                  `generated bench card ${current} of ${favoritePlants.length}`
-                );
-              } catch (e: any) {
-                if (e.message === "Download cancelled") {
-                  console.log(e.message);
-                } else {
-                  console.error(e);
-                }
-              }
-              return [p, b] as [Plant, Blob];
-            })
-          );
-          if (zipCancelled.current) {
-            throw new Error("Download cancelled");
-          }
-          var zip = new JSZip();
-          for (let [p, blob] of plantBlobPairs) {
-            zip.file(
-              p.commonName + ".pdf",
-              blob as unknown as null,
-              {
-                blob: true,
-              } as unknown as JSZip.JSZipFileOptions & { dir: true }
-            );
-          }
-          await zip.generateAsync({ type: "blob" }).then(function (content) {
-            saveAs(content, `bench-cards-${currentBct.name}.zip`);
-          });
-        } catch (e: any) {
-          if (e.message === "Download cancelled") {
-            console.log(e.message);
-          } else {
-            console.error(e);
-          }
-        } finally {
-          setShowZipModal(false);
-          setZipCurrent(0);
-          setZipTotal(0);
-          setCurrentBct(null);
-          zipCancelled.current = false;
-        }
-      };
-      cb();
-    }
-  }, [showZipModal, setShowZipModal, currentBct]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  const plantsAsExcelSheet = (
-    data: Data,
-    plants: Plant[],
-    regionNumbers: number[]
-  ): ExcelSheetData[] => 
-  [
-    {
-      columns: [
-        "Type(s)",
-        "Botanical Name",
-        "Common Name",
-        ...regionNumbers.flatMap(r => 
-          [
-            `Region ${r} Water Use`,
-            `Region ${r} ET0`,
-            `Region ${r} Plant Factor`
-          ])
-      ],
-      data: plants.map(
-        (p) =>
-          [
-            p.types.map((t) => data.plantTypeNameByCode[t]).join(", "),
-            p.botanicalName,
-            p.commonName,
-            ...regionNumbers.flatMap(r => 
-              [
-                data.waterUseByCode[p.waterUseByRegion[r-1]].name,
-                data.waterUseByCode[p.waterUseByRegion[r-1]].percentageET0 + "%",
-                data.waterUseByCode[p.waterUseByRegion[r-1]].plantFactor,
-              ])
-          ] as ExcelCellData[]
-      ),
-    } as ExcelSheetData
-  ];
-
-  const getDownloadActions = (
-    data: Data,
-    searchCriteria: SearchCriteria,
-    favoritePlants: Plant[]
-  ): DownloadAction[] => {
-    const ExcelFile = ReactExport.ExcelFile;
-    const ExcelSheet = ReactExport.ExcelFile.ExcelSheet;
-    const sideRender = (content: any) => {
-      let container = document.getElementById("download-outlet");
-      ReactDOM.render(<></>, container, () => {
-        /* IMPORTANT!  We clear the dom first, in order to force re-render.
-        ** Without this, the user can only download an excel document ONCE until they reload the page.  */
-        ReactDOM.render(content, container);
-      });
-    };
-    return [
-      {
-        include: !!searchCriteria.city,
-        method: () => {
-          Promise.all(
-            favoritePlants.map((p: Plant) =>
-              fetch(plantDetailQrCodeFromId(p.id).image_url)
-                .then((r) => r.blob())
-                .then((b) => [p, b] as [Plant, Blob])
-            )
-          ).then((plantBlobPairs) => {
-            var zip = new JSZip();
-            for (let [p, blob] of plantBlobPairs) {
-              zip.file(
-                p.commonName + ".png",
-                blob as unknown as null,
-                {
-                  blob: true,
-                } as unknown as JSZip.JSZipFileOptions & { dir: true }
-              );
-            }
-            zip.generateAsync({ type: "blob" }).then(function (content) {
-              saveAs(content, "qr-codes.zip");
-            });
-          });
-        },
-        label: (
-          <>
-            <FontAwesomeIcon icon={faQrcode} className="mr-2" />
-            Download QR codes
-          </>
-        ),
-      },
-      ...data.benchCardTemplates.map((bct) => ({
-        include: !!searchCriteria.city,
-        method: () => {
-          if (!zipCancelled.current) {
-            setCurrentBct(bct);
-            setShowZipModal(true);
-          }
-        },
-        label: (
-          <>
-            <FontAwesomeIcon icon={faIdCard} className="mr-2" />
-            Download Bench Cards ({bct.name})
-          </>
-        ),
-      })),
-      {
-        include: !!searchCriteria.city,
-        method: () => {
-          let excelData = plantsAsExcelSheet(
-            data,
-            favoritePlants,
-            [searchCriteria.city.region]
-          );
-          sideRender(
-            <ExcelFile
-              filename={`WUCOLS_${searchCriteria.city.name}`}
-              hideElement={true}
-            >
-              <ExcelSheet
-                dataSet={excelData}
-                name={`WUCOLS_${searchCriteria.city.name}`}
-              />
-            </ExcelFile>
-          );
-        },
-        label: (
-          <>
-            <FontAwesomeIcon icon={faFileExcel} className="mr-2" />
-            Download in Excel format
-          </>
-        ),
-      },
-      {
-        include: !searchCriteria.city,
-        method: () => {
-          console.log('preparing Excel...');
-          let excelData = plantsAsExcelSheet(
-            data,
-            data.plants,
-            [1,2,3,4,5,6]
-          );
-          console.log('rendering...');
-          sideRender(
-            <ExcelFile
-              filename={`WUCOLS_all_regions`}
-              hideElement={true}
-            >
-              <ExcelSheet
-                dataSet={excelData}
-                name={`WUCOLS_all_regions`}
-              />
-            </ExcelFile>
-          );
-          console.log('rendered');
-        },
-        label: (
-          <>
-            <FontAwesomeIcon icon={faFileExcel} className="mr-2" />
-            Download WUCOLS plants for all regions
-          </>
-        ),
-      }
-    ].filter(da => da.include);
-  };
 
   return (
     <>
@@ -606,7 +347,6 @@ function App({ data }: Props) {
             {...{
               favoritePlants,
               queryString: location.search,
-              getDownloadActions: getDownloadActions,
               isPlantFavorite,
               togglePlantFavorite,
               searchCriteria,
@@ -629,50 +369,12 @@ function App({ data }: Props) {
               togglePlantFavorite={togglePlantFavorite}
               setSearchCriteria={updateSearchCriteria}
               searchPerformed={searchWasPerformed(searchCriteria)}
-              downloadActions={getDownloadActions(data, searchCriteria, favoritePlants)}
               resetSearchCriteria={resetSearchCriteria}
               addAllToFavorites={addAllToFavorites}
               data={data}
             />
           )}
         />
-        <Modal
-          show={showZipModal}
-          onHide={() => {
-            zipCancelled.current = true;
-            setShowZipModal(false);
-          }}
-        >
-          <Modal.Header closeButton>
-            <Modal.Title>Generating Zip file</Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            <Container fluid>
-              <Row>
-                <Col>Please wait. It can take a few seconds per bech card.</Col>
-              </Row>
-              <Row>
-                <Col>
-                  <ProgressBar
-                    now={Math.round((zipCurrent / zipTotal) * 100)}
-                    label={`${zipCurrent} of ${zipTotal}`}
-                  />
-                </Col>
-              </Row>
-            </Container>
-          </Modal.Body>
-          <Modal.Footer>
-            <Button
-              variant="primary"
-              onClick={() => {
-                zipCancelled.current = true;
-                setShowZipModal(false);
-              }}
-            >
-              Cancel
-            </Button>
-          </Modal.Footer>
-        </Modal>
       </div>
     </>
   );
